Use Serde for handling JSON.
This patch replaces rustc-serialize with serde_json for JSON
serialization and deserialization. Most of this was already supported by
the build script changes in #149, but because the DynamoDB and KMS
helpers include automatic deserialization, they have been moved to
their own `*.in.rs` source files which are built by Serde and included
in the main `.rs` file for the appropriate module.

rustc-serialize can't be removed as a dependency completely because it's
still used for Base64 encoding and decoding. There is no Serde crate for
Base64 (yet).

I also updated the tests to remove rustc-serialize and to get rid of
some glob imports and added comments about KMS to the README that were
missing.

This closes #146.