Convert first character of locationName to uppercase when serializing.
Per botocore issue 1035, the EC2 query protocol requires an uppercase starting letter for locationName but the botocore model locationName starts with a lowercase letter.

Created a function to capitalize the first character of a string. This is used in generate_struct_field_serializers to generate correctly serialized names.

Fixes #383 and #387.