fix: make parsing test generation case-insensitive
Add `case_insensitive_btreemap_get` function to allow for case-insensitive
retrievals from `String`-keyed BTreeMap`s. This is necessary because of the way
`Operation` actions are determined for auto-generated tests: the action is
determined from the filename, which is lowercase, so there is no way to
determine the original casing in ambiguous scenarios. For example, 'mfa' should
be 'MFA', but there's no way for the codegen to know that, short of explicitly
hard-coding that into the test generation.