update to serde 1.0
fixes #621

this updates serde/serde_json/serde_derive to their 1.0 versions.
most of this work was just adding the appropriate lifetimes that
needed to be added for Deserialize (since in 1.0 they require lifetimes).

however when building I noticed I ran into failures with `min`/`max`,
on the `EMR`, and `Inspector` generation code. it turns out these
are actually floats, and we were parsing them as integers. so this
changes them to their float based values.