Add a streaming struct for shapes that might be used for streaming
This allows us to more cleanly generate output objects with the
appropriate type, without polluting all the code with logic to choose
between Vec<u8> or Box<Read>.

This also allows us to implement a dummy Debug trait on these types,
allowing the containing structures to inherit Debug as well. We need to
implement this as a struct (and not just a type alias) in order to be
able to implement the Debug trait (traits can only be implemented on a
type if either the type or trait is local to the crate).

We also implement Deref and DerefMut into Box<Read> so all the calling
code can easily just use Read trait functions.