Use streaming shapes for output fields that are meant to be streamed
This replaces primarily the Body type with a StreamingBody when we want
to perform operations that result in a stream of data to be read (such
as GetObject on S3). This allows us to read large blobs of data in a
streaming fashion rather than have it all turn up in one big chunk.