[wip] enable SCCache for builds
refs #774

what this pr does:

this pr adds in sccache which is a caching utility for rust builds.
this should really help take down our build times because almost all
of our time is spent building crates. by caching builds that haven't
changed with sccache we'll be able to greatly increase the speed of
travis/appveyor.

this is just one proposed solution to solving build time outs on PRs.
we'll need lots of discussion as to whether this is the right move,
and other alternatives :)

---

short explnantion of why this is wip:

this commit adds in [SCCache][sccache_link] for all CI Builds
(AppVeyor, and Travis alike). or I should say it would. I need
some help from maintainers to truly test, and see if this will
end up fixing our build woes for real, instead of just me
testing locally. a more winded explanation is below.

i know it seems like a lot of setup, but really if we don't like
sccache after testing we can undo all of this really easy, and
just attack the problem elsewhere.

---

long explnantion of why this is wip:

testing locally on my machine is great and all (infact I'm seeing
multiple minutes shaved off my build for just the service crates)
however lets be honest there's always a difference between local vs
travis vs appveyor.

however, in order to truly test this I need some help from the rusoto
maintainers :) those who can access encrypt variables in travis, and
those who have access to the AWS Account where rusoto will do it's
integration tests.

Specifically SCCache will need access to an S3 Bucket, and a pair
of encrypted AWS_ACCESS_KEY_ID/AWS_SECRET_ACCESS_KEY to access them.
Then we'll need someone to build SCCache for all platforms, and upload
them to some place public we trust (perhaps another S3 bucket in the same
account?). so that way we can download them in CI (building sccache is slow).

I've listed steps below for what would be the easiest way to perform this, and
how we can get it in this PR to test :) .

---

Things that I need help on getting ready:

- Building SCCache for All Platforms.

  The first thing we'll need to do is build sccache for all platforms
  so they can be downloaded later in CI. luckily sccache provides the
  ability to do this through a `build_release.sh` script in the main
  repo. I've included a link to it: [HERE][sccache_build_script].
  that being said I happen to have a mac/windows/linux box so if you
  need someone to help build I'd be happy to :) .

  Once SCCache has been built for all platforms we'll need to upload the
  resulting binaries somewhere that appveyor/travis can access them without
  auth. I personally recommend a public S3 bucket inside the AWS Testing account
  (this is how the core rust repo does it as well).

- Creating an S3 Bucket for SCCache + AWS Credentials

  We need to create an S3 Bucket for Rusoto to use SCCache, that it can use to store
  artifacts. We'll also need a user who can access this bucket (and pass in the values
  to sccache). the steps for this will look like:

    - Create an S3 Bucket in the rusoto AWS Account.
    - Create a user with `s3:*` on resources: `bucket-arn` && `bucket-arn/*`
    - Create an Access Key Pair/Secret Access Key for user that you just created.
    - Follow [travis guides][travis_encrypt] to encrypt Secret Access Key, and share
      the access key id + encrypted secret access key either in the PR comment, or
      by emailing me.
    - Follow [appveyor guides][appveyor_encrypt] to encrypt Secret Access Key, and share
      access key id + encrypted secret acacess key either in the PR comment, or by emailing
      me.

[sccache_link]: https://github.com/mozilla/sccache
[sccache_build_script]: https://github.com/mozilla/sccache/blob/master/scripts/build-release.sh
[travis_encrypt]: https://docs.travis-ci.com/user/encryption-keys#Usage
[appveyor_encrypt]: https://www.appveyor.com/docs/build-configuration/#secure-variables