fix uri encoding for services that use qs params
fixes #821

this fixes uri encoding for all services that end up using query
string parameters. the root cause of this issue is actual caused
by fixing our core signature generation.

you see according to multiple RFCS (that I'll point you to the issue
for) query strings are supposed to have '+'s represent ' ', and as
such be encoded as '%20'. This can be seen if you try to encode a
query string with a '+' as '%2B' when using V4 Auth and making a
request to S3.

However, sometimes we actually want a plus. This happens a lot, specifically
with our rust code. we get a return value from AWS, and we want to
physically use that '+' that AWS Returns.

There are two things that first come to mind:
  1. Make all of our signature code be techincally broken, not properly
    handling '+'s inside of query strings.

    - This is super unfortunate. Having funademently broken signing
      is really bad idea.

  2. Encode all of our rust fields inside of params.put() (so we put
    in an already encoded value).

    - This is unfortunate because it means if you actually want the
      signing behavior to handle the encoding of "+"'s for you when
      using the rust types, then you'll be thoroughly dissapointed.

Personally I think #2 is the much better solution. If you actually want
a space, pass in a space. So that's what I went with.

Side note:
  - If you notice I simply am using a "replace()", and not fully
    url encoding it. This has a couple benifits:
      - we don't have to pay the full cost of encoding twice.
        performance win yay?
      - don't have to have any extra imports.

All Integration tests should now pass (though specifically SQS).