Centralize struct attribute generation
This means that Clone is now derived for all shapes, with the exception
of streaming shapes, and those shapes which contain them.

Note this is functionally equivalent to the existing codegen except for
the three rest-xml services (cloudfront, route53 and s3), which will now
derive Clone where they did not before. All other services were already
deriving Clone for all shapes.