Remove Display implementation for Region
It is very easy to accidentally use <Region as Display>::fmt()
instead of Region::name() if the name is desired. In fact both
methods return the same unless a Region::Custom endpoint is in use.
So far I've seen this used incorrectly twice [1,2].

Alternatively, the Display implementation could be kept and `Region::name()`
could be removed. I prefer `Region::name()` though. It makes very clear that
the name is what you get.

[1]: https://github.com/rusoto/rusoto/pull/1018#pullrequestreview-113346404
[2]: 3033dea24a6cc23c8072d5f1baae0e6e9fc31162