Add support for prefixes for cred variables.
This fixes #1050, by providing a `with_prefix(...)` constructor as well as a
`Default` implementation for `EnvironmentProvider`. This is a breaking change
as a constructor will need to be used rather than invoking `credentials()` on
`EnvironmentProvider` directly, i.e.
`EnvironmentProvider::default().credentials().wait()`.

The `ChainProvider` has been updated to use the default constructor rather than
allowing for a custom `EnvironmentProvider` to be passed in. This can be
improved by migrating the `ChainProvider` construction into a Builder pattern.