Replace deprecated tokio_timer::Deadline
tokio-rs/tokio#558 replaced `Deadline` with `Timeout`, and deprecated
the former. Since it was released in the minor release
[0.2.6](https://github.com/tokio-rs/tokio/blob/master/tokio-timer/CHANGELOG.md#026-august-23-2018),
and rusoto doesn't check in `Cargo.lock`, but treats warnings as errors,
this prevents compilation when compiling in newer checkouts. We should
probably change this to avoid more breakage in the future, but for now,
this patch replaces all uses of `Deadline` with `Timeout`.