Run integration tests with `--test-threads 1`
Each test is starting its own Tokio runtime, which I don't expect people
to do in common use (outside of tests).

We're hitting a panic in hyper during the S3 integration tests, and we
need to do a bit of a refactor of the client to get around it, but let's
just run tests slow so we can get a beta out. :)