Ignore dead_code warnings on deserializers
We made a number of deserializer structs unnecessary in #1728 by simply
using the Default trait for empty shapes, but some attempts to remove
these structs reveal that a few of them are still in use.

This marks deserializers, which aren't `pub`, with `#[allow(dead_code)]`
to work around the complexity but still silence the warnings.