Make makefile more accessible
This is done by adding 'a magic line' which allows to add titles
to group targets, and doc strings for each target.

While running each and every target, I also noticed that service_codegen
depends on a git submodule being initialized, which is now happening
automatically when needed.

Furthermore, the `RUST_VERSION` make variable was never used within
make and could be removed.