Take reference to query input.
I have not yet used the other apis, but I will soon and probably the same thing applies to all of them. I don't see why queryinput needs to be taken by value when you only use a reference to it during json serialization. 

If you agree to this, is this the right place to change the source? It seems generated so probably somewhere else.