Added new_with_size to ByteStream
S3's PutObject doesn't work if there isn't a content-length, for
ByteStream's that length comes from size_hint, previously this was set
to None in the case of an unsized stream, which then made it unusable
for PutObject. Since there are many cases where this would be useful for
putting to S3, this commit adds a new_with_size constructor that takes a
size_hint.