Wake immediately before returning Poll::Pending
The original code returned Poll::Pending without first receiving
Poll::Pending from a downstream future or telling the task's waker
to wake the future immediately. This results in the future never
being called ever again (as nothing is told to wake it.

The original intent was for this future to be polled immediately
so more data would be read from the earlier part of poll_next().
Therefore, tell the task to wake this future immediately.