remove regex dependency
The dependency on `regex` showed up in the `cargo bloat` profile of a
large application I work on, and `rusoto` is the only crate which
depends on it. It appears that `rustoto`'s use of it is very light, so
this PR removes the regex dependency and replaces it with some simple
string splitting logic.